h4_21

+++ controllers/blogs.js
@@ -53,10 +53,24 @@ blogsRouter.post('/', async (request, response) => {
 blogsRouter.delete('/:id', async (request, response) => {
   try {
 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!decodedToken.id || !blog) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'unauthorized' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 
   } catch (exception) {
+    if (exception.name === 'JsonWebTokenError' ) {
+      return response.status(401).json({ error: exception.message })
+    }
 
     console.log(exception)
     response.status(400).send({ error: 'malformatted id' })
